Fix build error for the 32 bit build in the CI

+++ examples/tutorial_server_reverseconnect.c
@@ -17,7 +17,8 @@ static void stopHandler(int sig) {
 static void reverseConnectStateCallback(UA_Server *server, UA_UInt64 handle,
                                         UA_SecureChannelState state, void *context) {
     UA_LOG_INFO(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND,
-                "Reverse connect state callback for %lu with context %p: State %d", handle, context,  state);
+                "Reverse connect state callback for %lu with context %p: State %d",
+                (unsigned long)handle, context,  state);
 }
